+++ lib/email_processing/digest_email_parser.dart
@@ -7,7 +7,7 @@ import 'package:summer2022/models/MailResponse.dart';
 import 'package:summer2022/image_processing/google_cloud_vision_api.dart';
 import 'package:summer2022/models/Digest.dart';
 import 'package:summer2022/models/Code.dart';
-import 'image_processing/barcode_scanner.dart';
+import 'package:summer2022/image_processing/barcode_scanner.dart';
 import 'package:path_provider/path_provider.dart';
 import 'package:permission_handler/permission_handler.dart';
 import 'package:summer2022/image_processing/usps_address_verification.dart';
