+++ lib/email_processing/other_mail_parser.dart
@@ -1,6 +1,6 @@
 import 'package:enough_mail/enough_mail.dart';
 import 'package:intl/intl.dart';
-import 'models/Digest.dart';
+import 'package:summer2022/models/Digest.dart';
 
 class OtherMailParser {
   String _userName = ''; // Add your credentials
