+++ lib/ui/main_menu.dart
@@ -6,12 +6,12 @@ import 'package:summer2022/image_processing/imageProcessing.dart';
 import 'package:flutter/material.dart';
 import 'package:image_picker/image_picker.dart';
 import 'package:loader_overlay/loader_overlay.dart';
-import 'package:summer2022/digest_email_parser.dart';
-import 'package:summer2022/other_mail_parser.dart';
-import 'package:summer2022/read_info.dart';
-import 'package:summer2022/Keychain.dart';
+import 'package:summer2022/email_processing/digest_email_parser.dart';
+import 'package:summer2022/email_processing/other_mail_parser.dart';
+import 'package:summer2022/speech_commands/read_info.dart';
+import 'package:summer2022/utility/Keychain.dart';
 import 'package:summer2022/image_processing/google_cloud_vision_api.dart';
-import 'package:summer2022/main.dart';
+import 'package:summer2022/ui/main.dart';
 import 'package:summer2022/models/Arguments.dart';
 import 'package:summer2022/models/EmailArguments.dart';
 import 'package:summer2022/models/Digest.dart';
