+++ lib/ui/sign_in.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:summer2022/Keychain.dart';
-import 'package:summer2022/Client.dart';
-import 'package:summer2022/main.dart';
+import 'package:summer2022/utility/Keychain.dart';
+import 'package:summer2022/utility/Client.dart';
+import 'package:summer2022/ui/main.dart';
 import 'package:summer2022/ui/bottom_app_bar.dart';
 import 'package:loader_overlay/loader_overlay.dart';
