+++ lib/utility/RouteGenerator.dart
@@ -1,12 +1,12 @@
 import 'package:flutter/material.dart';
-import 'ui/mail_widget.dart';
-import 'ui/main_menu.dart';
-import 'ui/other_mail.dart';
-import 'ui/settings.dart';
-import 'ui/sign_in.dart';
-import './backend_testing.dart';
-import 'models/Arguments.dart';
-import 'models/EmailArguments.dart';
+import 'package:summer2022/ui/mail_widget.dart';
+import 'package:summer2022/ui/main_menu.dart';
+import 'package:summer2022/ui/other_mail.dart';
+import 'package:summer2022/ui/settings.dart';
+import 'package:summer2022/ui/sign_in.dart';
+import 'package:summer2022/backend_testing.dart';
+import 'package:summer2022/models/Arguments.dart';
+import 'package:summer2022/models/EmailArguments.dart';
 
 class RouteGenerator {
   static Route<dynamic> generateRoute(RouteSettings settings) {
