+++ test/command_tutorial_test.dart
@@ -1,8 +1,8 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:flutter_tts/flutter_tts.dart';
 import 'package:global_configuration/global_configuration.dart';
-import 'package:summer2022/read_info.dart';
-import 'package:summer2022/main.dart';
+import 'package:summer2022/speech_commands/read_info.dart';
+import 'package:summer2022/ui/main.dart';
 
 
 class FakeFlutterTts extends Fake implements FlutterTts {
