--- test/data/write_to_file_test.json
@@ -1,1 +1,0 @@
-{"addresses":[{"type":"sender","name":"GEICO","address":"2563 Forest Dr; Annapolis, MD 21401","validated":false},{"type":"recipient","name":"Deborah Keenan","address":"1006 Morgan Station Dr; Severn, MD 21144-1245","validated":true}],"logos":[{"name":"GEICO"},{"name":"GEICO"}],"codes":[{"type":"url","info":"https://geico.com"}]}
