+++ test/speech_to_text_test.dart
@@ -1,5 +1,5 @@
 import 'package:flutter_test/flutter_test.dart';
-import 'package:summer2022/speech_to_text.dart';
+import 'package:summer2022/speech_commands/speech_to_text.dart';
 import 'dart:io' as io;
 
 void main() async {
